Refactor to modular architecture
