+++ TheBackendCmsSolution/Modules/Content/Dtos/ContentTypeDto.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace TheBackendCmsSolution.ApiService.Dtos
+namespace TheBackendCmsSolution.Modules.Content.Dtos
 {
     public class ContentTypeDto
     {
