+++ TheBackendCmsSolution/Modules/Taxonomy/TheBackendCmsSolution.Modules.Taxonomy.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="../Abstractions/TheBackendCmsSolution.Modules.Abstractions.csproj" />
+  </ItemGroup>
+</Project>
