+++ TheBackendCmsSolution/TheBackendCmsSolution.ApiService/TheBackendCmsSolution.ApiService.csproj
@@ -7,19 +7,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.0.6">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.4" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\TheBackendCmsSolution.ServiceDefaults\TheBackendCmsSolution.ServiceDefaults.csproj" />
+    <ProjectReference Include="..\Modules\Abstractions\TheBackendCmsSolution.Modules.Abstractions.csproj" />
+    <ProjectReference Include="..\Modules\Content\TheBackendCmsSolution.Modules.Content.csproj" />
+    <ProjectReference Include="..\Modules\Users\TheBackendCmsSolution.Modules.Users.csproj" />
+    <ProjectReference Include="..\Modules\Taxonomy\TheBackendCmsSolution.Modules.Taxonomy.csproj" />
   </ItemGroup>
 
 </Project>
