+++ TheBackendCmsSolution/TheBackendCmsSolution.sln
@@ -10,6 +10,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheBackendCmsSolution.ApiSe
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheBackendCmsSolution.Web", "TheBackendCmsSolution.Web\TheBackendCmsSolution.Web.csproj", "{2B6A07B0-9971-4196-A550-EC9DE359262E}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheBackendCmsSolution.Modules.Abstractions", "Modules\Abstractions\TheBackendCmsSolution.Modules.Abstractions.csproj", "{52F401ED-A5DA-4EA2-838C-4A73B9D93B49}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheBackendCmsSolution.Modules.Content", "Modules\Content\TheBackendCmsSolution.Modules.Content.csproj", "{0E63E9EF-7372-4CCA-BD89-79C849D99955}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheBackendCmsSolution.Modules.Users", "Modules\Users\TheBackendCmsSolution.Modules.Users.csproj", "{E68BCFC7-466E-41D4-A028-F49B8885773E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TheBackendCmsSolution.Modules.Taxonomy", "Modules\Taxonomy\TheBackendCmsSolution.Modules.Taxonomy.csproj", "{3BD5038F-3405-4BD3-BED5-A28A00E00234}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,10 +37,26 @@ Global
 		{AEE088E8-C707-4783-B079-E31B39633A84}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AEE088E8-C707-4783-B079-E31B39633A84}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2B6A07B0-9971-4196-A550-EC9DE359262E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B6A07B0-9971-4196-A550-EC9DE359262E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2B6A07B0-9971-4196-A550-EC9DE359262E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2B6A07B0-9971-4196-A550-EC9DE359262E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
+                {2B6A07B0-9971-4196-A550-EC9DE359262E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {2B6A07B0-9971-4196-A550-EC9DE359262E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {2B6A07B0-9971-4196-A550-EC9DE359262E}.Release|Any CPU.Build.0 = Release|Any CPU
+                {52F401ED-A5DA-4EA2-838C-4A73B9D93B49}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {52F401ED-A5DA-4EA2-838C-4A73B9D93B49}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {52F401ED-A5DA-4EA2-838C-4A73B9D93B49}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {52F401ED-A5DA-4EA2-838C-4A73B9D93B49}.Release|Any CPU.Build.0 = Release|Any CPU
+                {0E63E9EF-7372-4CCA-BD89-79C849D99955}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {0E63E9EF-7372-4CCA-BD89-79C849D99955}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {0E63E9EF-7372-4CCA-BD89-79C849D99955}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {0E63E9EF-7372-4CCA-BD89-79C849D99955}.Release|Any CPU.Build.0 = Release|Any CPU
+                {E68BCFC7-466E-41D4-A028-F49B8885773E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {E68BCFC7-466E-41D4-A028-F49B8885773E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {E68BCFC7-466E-41D4-A028-F49B8885773E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {E68BCFC7-466E-41D4-A028-F49B8885773E}.Release|Any CPU.Build.0 = Release|Any CPU
+                {3BD5038F-3405-4BD3-BED5-A28A00E00234}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {3BD5038F-3405-4BD3-BED5-A28A00E00234}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {3BD5038F-3405-4BD3-BED5-A28A00E00234}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {3BD5038F-3405-4BD3-BED5-A28A00E00234}.Release|Any CPU.Build.0 = Release|Any CPU
+        EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
